+++ src/controllers/user.controller.ts
@@ -1,13 +1,7 @@
 import httpStatusCodes from 'http-status-codes';
 import logger from '../config/logger';
-import {
-  ERRORS,
-  MESSAGE_ERRORS,
-  RESPONSE_ERROR,
-  SOCIAL_ERRORS,
-} from '../constants/errors';
-import { MESSAGE_RESPONSE, USER_RESPONSE } from '../constants/successMessages';
-import Utility from '../constants/utility';
+import { ERRORS, RESPONSE_ERROR, SOCIAL_ERRORS } from '../constants/errors';
+import { USER_RESPONSE } from '../constants/successMessages';
 import UserService from '../services/user.service';
 import apiResponse from '../utilities/apiResponse';
 
@@ -211,206 +205,6 @@ export class UserController {
           message: RESPONSE_ERROR.RES_ERROR,
         });
       }
-    }
-  }
-
-  //===================================== MESSAGE ====================================
-
-  public static async sendMessage(req, res) {
-    const { accountId } = req.params; //receiverId
-    const { user } = req; //sender
-    const { newMessage } = req.body;
-
-    try {
-      const sentMessage = await UserService.sendMessage(
-        accountId,
-        user.accountId,
-        newMessage
-      );
-      return apiResponse.result(
-        res,
-        {
-          message: MESSAGE_RESPONSE.MESSAGE_CREATE,
-          sentMessage,
-        },
-        httpStatusCodes.CREATED
-      );
-    } catch (e) {
-      logger.error('[userController.sendMessage]:' + e.message);
-      return Utility.apiErrorResponse(res, e, [
-        ERRORS.USER_DOES_NOT_EXIST,
-        MESSAGE_ERRORS.PRIVATE_USER,
-      ]);
-    }
-  }
-
-  public static async sendGroupMessage(req, res) {
-    const { chatGroupId } = req.params; //chat group Id
-    const { user } = req; //sender
-    const { newMessage } = req.body;
-
-    try {
-      const sentMessage = await UserService.sendGroupMessage(
-        chatGroupId,
-        user.accountId,
-        newMessage
-      );
-      return apiResponse.result(
-        res,
-        {
-          message: MESSAGE_RESPONSE.MESSAGE_CREATE,
-          sentMessage,
-        },
-        httpStatusCodes.CREATED
-      );
-    } catch (e) {
-      logger.error('[userController.sendGroupMessage]:' + e.message);
-      return Utility.apiErrorResponse(res, e, [
-        ERRORS.USER_DOES_NOT_EXIST,
-        MESSAGE_ERRORS.CHAT_GROUP_MISSING,
-      ]);
-    }
-  }
-
-  public static async getAllMessages(req, res) {
-    const { accountId } = req.params;
-    const { user } = req;
-    try {
-      const listOfMessages = await UserService.getAllMessages(
-        accountId,
-        user.accountId
-      );
-      return apiResponse.result(
-        res,
-        {
-          message: 'success',
-          listOfMessages,
-        },
-        httpStatusCodes.OK
-      );
-    } catch (e) {
-      logger.error('[userController.getAllMessages]:' + e.message);
-      return apiResponse.error(res, httpStatusCodes.INTERNAL_SERVER_ERROR, {
-        message: RESPONSE_ERROR.RES_ERROR,
-      });
-    }
-  }
-
-  public static async getChatList(req, res) {
-    const { accountId } = req.params;
-
-    try {
-      const chatList = await UserService.getChatList(accountId);
-      return apiResponse.result(
-        res,
-        {
-          message: 'success',
-          chatList,
-        },
-        httpStatusCodes.OK
-      );
-    } catch (e) {
-      logger.error('[userController.getChatList]:' + e.message);
-      return apiResponse.error(res, httpStatusCodes.INTERNAL_SERVER_ERROR, {
-        message: RESPONSE_ERROR.RES_ERROR,
-      });
-    }
-  }
-
-  public static async createChatGroup(req, res) {
-    const { user } = req; //creator of chat group
-    const { newChatGroup } = req.body;
-
-    try {
-      const groupCreated = await UserService.createChatGroup(
-        user.accountId,
-        newChatGroup
-      );
-      return apiResponse.result(
-        res,
-        {
-          message: MESSAGE_RESPONSE.GROUP_CREATE,
-          groupCreated,
-        },
-        httpStatusCodes.CREATED
-      );
-    } catch (e) {
-      logger.error('[userController.createChatGroup]:' + e.message);
-      return Utility.apiErrorResponse(res, e, [ERRORS.USER_DOES_NOT_EXIST]);
-    }
-  }
-  public static async addUserToChatGroup(req, res) {
-    const { user } = req;
-    const { chatGroupId, accountId } = req.params;
-
-    try {
-      const userAdded = await UserService.addUserToChatGroup(
-        user.accountId,
-        accountId,
-        chatGroupId
-      );
-      return apiResponse.result(
-        res,
-        {
-          message: MESSAGE_RESPONSE.USER_ADDED,
-          userAdded,
-        },
-        httpStatusCodes.CREATED
-      );
-    } catch (e) {
-      logger.error('[userController.addUserToChatGroup]:' + e.message);
-      return Utility.apiErrorResponse(res, e, [
-        ERRORS.USER_DOES_NOT_EXIST,
-        httpStatusCodes.getStatusText(httpStatusCodes.UNAUTHORIZED),
-        MESSAGE_ERRORS.USER_CANNOT_BE_ADDED,
-        MESSAGE_ERRORS.CHAT_GROUP_MISSING,
-      ]);
-    }
-  }
-
-  public static async removeUserFromChatGroup(req, res) {
-    const { user } = req;
-    const { chatGroupId, accountId } = req.params;
-
-    try {
-      await UserService.removeUserFromChatGroup(
-        user.accountId,
-        accountId,
-        chatGroupId
-      );
-      return apiResponse.result(
-        res,
-        { message: MESSAGE_RESPONSE.USER_REMOVED },
-        httpStatusCodes.OK
-      );
-    } catch (e) {
-      logger.error('[userController.removeUserFromChatGroup]:' + e.message);
-      return Utility.apiErrorResponse(res, e, [
-        ERRORS.USER_DOES_NOT_EXIST,
-        httpStatusCodes.getStatusText(httpStatusCodes.UNAUTHORIZED),
-        MESSAGE_ERRORS.CHAT_GROUP_MISSING,
-      ]);
-    }
-  }
-
-  public static async deleteChatGroup(req, res) {
-    const { user } = req;
-    const { chatGroupId } = req.params;
-
-    try {
-      await UserService.deleteChatGroup(user.accountId, chatGroupId);
-      return apiResponse.result(
-        res,
-        { message: MESSAGE_RESPONSE.GROUP_DELETED },
-        httpStatusCodes.OK
-      );
-    } catch (e) {
-      logger.error('[userController.deleteChatGroup]:' + e.message);
-      return Utility.apiErrorResponse(res, e, [
-        ERRORS.USER_DOES_NOT_EXIST,
-        httpStatusCodes.getStatusText(httpStatusCodes.UNAUTHORIZED),
-        MESSAGE_ERRORS.CHAT_GROUP_MISSING,
-      ]);
     }
   }
 }
